fmtowns: fixed regression (no whatsnew)
git-svn-id: svn://messdev.no-ip.org/mess@14894 9849079f-6229-0410-a07f-84406ce305ba

+++ src/mess/includes/fmtowns.h
@@ -185,7 +185,7 @@ class towns_state : public driver_device
 	UINT8 m_prev_x;
 	UINT8 m_prev_y;
 
-	required_shared_ptr<UINT8> m_nvram;
+	required_shared_ptr<UINT32> m_nvram;
 
 	virtual void driver_start();
 	virtual void machine_reset();
